chore: adds small file size sound
